+++ src/ark/segmentation/marker_quantification.py
@@ -10,7 +10,7 @@ from alpineer import io_utils, load_utils, misc_utils
 import ark.settings as settings
 from ark.segmentation.regionprops_extraction import REGIONPROPS_FUNCTION
 from ark.segmentation.signal_extraction import EXTRACTION_FUNCTION
-from ark.utils import segmentation_utils
+from ark.segmentation import segmentation_utils
 
 
 def get_single_compartment_props(segmentation_labels, regionprops_base,
