+++ src/ark/utils/deepcell_service_utils.py
@@ -14,7 +14,7 @@ from requests.adapters import HTTPAdapter
 from requests.exceptions import RetryError
 from requests.packages.urllib3.util import Retry
 from tifffile import imread
-from alpineer import image_utils, io_utils, misc_utils
+from alpineer import image_utils, io_utils, misc_utils, load_utils
 from tqdm.notebook import tqdm
 
 
@@ -324,6 +324,71 @@ def run_deepcell_direct(input_dir, output_dir, host='https://deepcell.org',
     )
 
     return 0
+
+
+# TODO: Add metadata for channel name (eliminates need for fixed-order channels)
+def generate_deepcell_input(data_dir, tiff_dir, nuc_channels, mem_channels, fovs,
+                            is_mibitiff=False, img_sub_folder="TIFs", dtype="int16"):
+    """Saves nuclear and membrane channels into deepcell input format.
+    Either nuc_channels or mem_channels should be specified.
+
+    Writes summed channel images out as multitiffs (channels first).
+
+    Args:
+        data_dir (str):
+            location to save deepcell input tifs
+        tiff_dir (str):
+            directory containing folders of images, is_mibitiff determines what type
+        nuc_channels (list):
+            nuclear channels to be summed over
+        mem_channels (list):
+            membrane channels to be summed over
+        fovs (list):
+            list of folders to or MIBItiff files to load imgs from
+        is_mibitiff (bool):
+            if the images are of type MIBITiff
+        img_sub_folder (str):
+            if is_mibitiff is False, define the image subfolder for each fov
+            ignored if is_mibitiff is True
+        dtype (str/type):
+            optional specifier of image type.  Overwritten with warning for float images
+    Raises:
+        ValueError:
+            Raised if nuc_channels and mem_channels are both None or empty
+    """
+
+    # cannot have no nuclear and no membrane channels
+    if not nuc_channels and not mem_channels:
+        raise ValueError('Either nuc_channels or mem_channels should be non-empty.')
+
+    # define the channels list by combining nuc_channels and mem_channels
+    channels = (nuc_channels if nuc_channels else []) + (mem_channels if mem_channels else [])
+
+    # filter channels for None (just in case)
+    channels = [channel for channel in channels if channel is not None]
+
+    for fov in fovs:
+        # load the images in the current fov batch
+        if is_mibitiff:
+            data_xr = load_utils.load_imgs_from_mibitiff(
+                tiff_dir, mibitiff_files=[fov], channels=channels
+            )
+        else:
+            data_xr = load_utils.load_imgs_from_tree(
+                tiff_dir, img_sub_folder=img_sub_folder, fovs=[fov], channels=channels
+            )
+
+        fov_name = data_xr.fovs.values[0]
+        out = np.zeros((2, data_xr.shape[1], data_xr.shape[2]), dtype=data_xr.dtype)
+
+        # sum over channels and add to output
+        if nuc_channels:
+            out[0] = np.sum(data_xr.loc[fov_name, :, :, nuc_channels].values, axis=2)
+        if mem_channels:
+            out[1] = np.sum(data_xr.loc[fov_name, :, :, mem_channels].values, axis=2)
+
+        save_path = os.path.join(data_dir, f"{fov_name}.tiff")
+        image_utils.save_image(save_path, out)
 
 
 def _convert_deepcell_seg_masks(seg_mask: bytes) -> np.ndarray:
