+++ tests/analysis/spatial_analysis_test.py
@@ -10,7 +10,7 @@ from alpineer.test_utils import _write_labels
 
 import ark.settings as settings
 from ark.analysis import spatial_analysis
-from ark.utils import spatial_analysis_utils
+from ark.analysis import spatial_analysis_utils
 import test_utils
