+++ tests/segmentation/segmentation_utils_test.py
@@ -10,7 +10,7 @@ from alpineer import image_utils
 from alpineer.test_utils import make_images_xarray, make_labels_xarray
 
 import ark.settings as settings
-from ark.utils import segmentation_utils
+from ark.segmentation import segmentation_utils
 import test_utils
