+++ tests/utils/data_utils_test.py
@@ -479,117 +479,6 @@ def test_generate_and_save_neighborhood_cluster_masks(sub_dir, name_suffix):
             assert np.all(np.isin(neighborhood_mask, np.array([10 * i for i in np.arange(6)])))
 
 
-def test_generate_deepcell_input():
-    for is_mibitiff in [False, True]:
-        with tempfile.TemporaryDirectory() as temp_dir:
-            fovs = ['fov1', 'fov2', 'fov3']
-            chans = ['nuc1', 'nuc2', 'mem1', 'mem2']
-
-            tiff_dir = os.path.join(temp_dir, 'tiff_dir')
-            os.mkdir(tiff_dir)
-
-            if is_mibitiff:
-                fov_paths, data_xr = test_utils.create_paired_xarray_fovs(
-                    tiff_dir, fov_names=fovs, channel_names=chans, mode='mibitiff', dtype='int16'
-                )
-
-                # because we're matching files and not directories for mibitiffs
-                fovs = [fov + '.tiff' for fov in fovs]
-            else:
-                fov_paths, data_xr = test_utils.create_paired_xarray_fovs(
-                    tiff_dir, fov_names=fovs, channel_names=chans, dtype='int16', sub_dir='TIFs'
-                )
-
-            # test 1 nuc, 1 mem (no summing)
-            nucs = ['nuc2']
-            mems = ['mem2']
-
-            fov1path = os.path.join(temp_dir, 'fov1.tiff')
-            fov2path = os.path.join(temp_dir, 'fov2.tiff')
-            fov3path = os.path.join(temp_dir, 'fov3.tiff')
-
-            data_utils.generate_deepcell_input(
-                data_dir=temp_dir, tiff_dir=tiff_dir, nuc_channels=nucs, mem_channels=mems,
-                fovs=fovs, is_mibitiff=is_mibitiff, img_sub_folder='TIFs'
-            )
-
-            fov1 = np.moveaxis(io.imread(fov1path), 0, -1)
-            fov2 = np.moveaxis(io.imread(fov2path), 0, -1)
-            fov3 = np.moveaxis(io.imread(fov3path), 0, -1)
-
-            assert np.array_equal(fov1, data_xr.loc['fov1', :, :, ['nuc2', 'mem2']].values)
-            assert np.array_equal(fov2, data_xr.loc['fov2', :, :, ['nuc2', 'mem2']].values)
-            assert np.array_equal(fov3, data_xr.loc['fov3', :, :, ['nuc2', 'mem2']].values)
-
-            # test 2 nuc, 2 mem (summing)
-            nucs = ['nuc1', 'nuc2']
-            mems = ['mem1', 'mem2']
-
-            data_utils.generate_deepcell_input(
-                data_dir=temp_dir, tiff_dir=tiff_dir, nuc_channels=nucs, mem_channels=mems,
-                fovs=fovs, is_mibitiff=is_mibitiff, img_sub_folder='TIFs'
-            )
-
-            nuc_sums = data_xr.loc[:, :, :, nucs].sum(dim='channels').values
-            mem_sums = data_xr.loc[:, :, :, mems].sum(dim='channels').values
-
-            fov1 = np.moveaxis(io.imread(fov1path), 0, -1)
-            fov2 = np.moveaxis(io.imread(fov2path), 0, -1)
-            fov3 = np.moveaxis(io.imread(fov3path), 0, -1)
-
-            assert np.array_equal(fov1[:, :, 0], nuc_sums[0, :, :])
-            assert np.array_equal(fov1[:, :, 1], mem_sums[0, :, :])
-            assert np.array_equal(fov2[:, :, 0], nuc_sums[1, :, :])
-            assert np.array_equal(fov2[:, :, 1], mem_sums[1, :, :])
-            assert np.array_equal(fov3[:, :, 0], nuc_sums[2, :, :])
-            assert np.array_equal(fov3[:, :, 1], mem_sums[2, :, :])
-
-            # test nuc None
-            nucs = None
-
-            data_utils.generate_deepcell_input(
-                data_dir=temp_dir, tiff_dir=tiff_dir, nuc_channels=nucs, mem_channels=mems,
-                fovs=fovs, is_mibitiff=is_mibitiff, img_sub_folder='TIFs'
-            )
-
-            fov1 = np.moveaxis(io.imread(fov1path), 0, -1)
-            fov2 = np.moveaxis(io.imread(fov2path), 0, -1)
-            fov3 = np.moveaxis(io.imread(fov3path), 0, -1)
-
-            assert np.all(fov1[:, :, 0] == 0)
-            assert np.array_equal(fov1[:, :, 1], mem_sums[0, :, :])
-            assert np.all(fov2[:, :, 0] == 0)
-            assert np.array_equal(fov2[:, :, 1], mem_sums[1, :, :])
-            assert np.all(fov3[:, :, 0] == 0)
-            assert np.array_equal(fov3[:, :, 1], mem_sums[2, :, :])
-
-            # test mem None
-            nucs = ['nuc2']
-            mems = None
-
-            data_utils.generate_deepcell_input(
-                data_dir=temp_dir, tiff_dir=tiff_dir, nuc_channels=nucs, mem_channels=mems,
-                fovs=fovs, is_mibitiff=is_mibitiff, img_sub_folder='TIFs'
-            )
-
-            fov1 = np.moveaxis(io.imread(fov1path), 0, -1)
-            fov2 = np.moveaxis(io.imread(fov2path), 0, -1)
-            fov3 = np.moveaxis(io.imread(fov3path), 0, -1)
-
-            assert np.all(fov1[:, :, 1] == 0)
-            assert np.array_equal(fov1[:, :, 0], data_xr.loc['fov1', :, :, 'nuc2'].values)
-            assert np.all(fov2[:, :, 1] == 0)
-            assert np.array_equal(fov2[:, :, 0], data_xr.loc['fov2', :, :, 'nuc2'].values)
-            assert np.all(fov3[:, :, 1] == 0)
-            assert np.array_equal(fov3[:, :, 0], data_xr.loc['fov3', :, :, 'nuc2'].values)
-
-            # test nuc None and mem None
-            with pytest.raises(ValueError):
-                data_utils.generate_deepcell_input(
-                    data_xr, temp_dir, None, None, ['fov0'], ['chan0']
-                )
-
-
 def test_split_img_stack():
     with tempfile.TemporaryDirectory() as temp_dir:
         fovs = ['stack_sample']
